Conflicts:
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFacade.cxx
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFacade.h
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFromDirectory.cxx
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFromDirectory.h
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkStereoCameraCalibrationFromTwoDirectories.cxx
	Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkStereoCameraCalibrationFromTwoDirectories.h

+++ Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFacade.h
@@ -102,8 +102,8 @@ extern "C++" NIFTKOPENCV_EXPORT bool ExtractChessBoardPoints(
   const bool& drawCorners,
   const double& squareSizeInMillimetres,
   const mitk::Point2D& pixelScaleFactor,
-  std::vector<cv::Point2f>*& outputImagePoints,
-  std::vector<cv::Point3f>*& outputObjectPoints
+  std::vector<cv::Point2d>*& outputImagePoints,
+  std::vector<cv::Point3d>*& outputObjectPoints
   );
 
 
@@ -281,7 +281,7 @@ extern "C++" NIFTKOPENCV_EXPORT void OutputCalibrationData(
   const CvMat& distortionCoefficients,
   const CvMat& rotationVectors,
   const CvMat& translationVectors,
-  const float& projectionError,
+  const double& projectionError,
   const int& sizeX,
   const int& sizeY,
   const int& cornersX,
