+++ Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkCameraCalibrationFromDirectory.h
@@ -44,7 +44,7 @@ public:
   double Calibrate(const std::string& fullDirectoryName,
       const int& numberCornersX,
       const int& numberCornersY,
-      const float& sizeSquareMillimeters,
+      const double& sizeSquareMillimeters,
       const mitk::Point2D& pixelScaleFactor,
       const std::string& outputFile,
       const bool& writeImages
