+++ Code/Gui/MITK/Modules/niftkOpenCV/CameraCalibration/mitkStereoCameraCalibrationFromTwoDirectories.h
@@ -45,7 +45,7 @@ public:
       const std::string& rightDirectoryName,
       const int& numberCornersX,
       const int& numberCornersY,
-      const float& sizeSquareMillimeters,
+      const double& sizeSquareMillimeters,
       const mitk::Point2D& pixelScaleFactor,
       const std::string& outputFileName,
       const bool& writeImages
